organize settings elements
